Refactor saving/loading operations to Storage class
